pyproject.toml
[tool.poetry]
name = "django-rest-framework"
version = "0.1.0"
description = ""
authors = ["ludmila davletshina <ludaplv@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
coverage = "^7.6.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
